move products crud and fix imports
