+++ frontend/src/components/products/ProductsCrud/editableListProductData.module.scss
