+++ frontend/src/components/products/ProductsCrud/editableProductData.module.scss
